+++ features/manage-users/src/main/java/com/carissa/revibes/manage_users/data/model/CreateUserRequest.kt
@@ -29,6 +29,6 @@ data class CreateUserData(
     val email: String,
     val phoneNumber: String,
     val role: String,
-    val points: Int,
+    val points: Long,
     val createdAt: String,
 )
